requirements.txt
pywebview==5.1
pillow==10.4.0
python-dateutil==2.9.0.post0

pytesseract==0.3.10
soundfile==0.12.1

openai==1.61.1
transformers>=4.50.0
accelerate==0.34.2
sentencepiece==0.2.0

faster-whisper==1.1.0

google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1

icalendar==5.0.13
